# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fanfic Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ScrapeFicBody {
  url: string;
}

export interface ScrapedFic {
  url: string;
  title: string;
  author: string;
  fandom: string;
  /** @nullable */
  ship?: string | null;
  wordCount: number;
  tags: string[];
}

export interface Fic {
  id: string;
  url: string;
  title: string;
  author: string;
  fandom: string;
  /** @nullable */
  ship?: string | null;
  wordCount: number;
  tags: string[];
  dateAdded: string;
  /** @nullable */
  userRating?: number | null;
  /** @nullable */
  userNote?: string | null;
}

export interface CreateFicBody {
  url: string;
  title: string;
  author: string;
  fandom: string;
  /** @nullable */
  ship?: string | null;
  wordCount: number;
  tags: string[];
  /** @nullable */
  userRating?: number | null;
  /** @nullable */
  userNote?: string | null;
}

export interface UpdateFicBody {
  /** @nullable */
  userRating?: number | null;
  /** @nullable */
  userNote?: string | null;
}

export interface MonthlyStats {
  ficCount: number;
  fandomCount: number;
  month: string;
}