# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fanfic Tracker API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Scrape AO3 fic metadata from URL
 */
export const ScrapeFicBody = zod.object({
  url: zod.string(),
});

export const ScrapeFicResponse = zod.object({
  url: zod.string(),
  title: zod.string(),
  author: zod.string(),
  fandom: zod.string(),
  ship: zod.string().nullish(),
  wordCount: zod.number(),
  tags: zod.array(zod.string()),
});

/**
 * @summary List all logged fics sorted by date added descending
 */
export const ListFicsResponseItem = zod.object({
  id: zod.string(),
  url: zod.string(),
  title: zod.string(),
  author: zod.string(),
  fandom: zod.string(),
  ship: zod.string().nullish(),
  wordCount: zod.number(),
  tags: zod.array(zod.string()),
  dateAdded: zod.string(),
  userRating: zod.number().nullish(),
  userNote: zod.string().nullish(),
});
export const ListFicsResponse = zod.array(ListFicsResponseItem);

/**
 * @summary Log a new fic
 */
export const CreateFicBody = zod.object({
  url: zod.string(),
  title: zod.string(),
  author: zod.string(),
  fandom: zod.string(),
  ship: zod.string().nullish(),
  wordCount: zod.number(),
  tags: zod.array(zod.string()),
  userRating: zod.number().nullish(),
  userNote: zod.string().nullish(),
});

/**
 * @summary Get a single fic record
 */
export const GetFicParams = zod.object({
  id: zod.coerce.string(),
});

export const GetFicResponse = zod.object({
  id: zod.string(),
  url: zod.string(),
  title: zod.string(),
  author: zod.string(),
  fandom: zod.string(),
  ship: zod.string().nullish(),
  wordCount: zod.number(),
  tags: zod.array(zod.string()),
  dateAdded: zod.string(),
  userRating: zod.number().nullish(),
  userNote: zod.string().nullish(),
});

/**
 * @summary Update fic rating and note
 */
export const UpdateFicParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateFicBody = zod.object({
  userRating: zod.number().nullish(),
  userNote: zod.string().nullish(),
});

export const UpdateFicResponse = zod.object({
  id: zod.string(),
  url: zod.string(),
  title: zod.string(),
  author: zod.string(),
  fandom: zod.string(),
  ship: zod.string().nullish(),
  wordCount: zod.number(),
  tags: zod.array(zod.string()),
  dateAdded: zod.string(),
  userRating: zod.number().nullish(),
  userNote: zod.string().nullish(),
});

/**
 * @summary Delete a fic record
 */
export const DeleteFicParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Get fic count and fandom count for the current month
 */
export const GetMonthlyStatsResponse = zod.object({
  ficCount: zod.number(),
  fandomCount: zod.number(),
  month: zod.string(),
});