# lib/api-zod/src/generated/types/fic.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fanfic Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Fic {
  id: string;
  url: string;
  title: string;
  author: string;
  fandom: string;
  /** @nullable */
  ship?: string | null;
  wordCount: number;
  tags: string[];
  dateAdded: string;
  /** @nullable */
  userRating?: number | null;
  /** @nullable */
  userNote?: string | null;
}