# lib/api-zod/src/generated/types/monthlyStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fanfic Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyStats {
  ficCount: number;
  fandomCount: number;
  month: string;
}